refactor(step2): 서비스 레이어 제거
